backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
pgvector
sentence-transformers
torch
requests
djangorestframework-simplejwt
djoser